Code with now 40sec exectution time instead of 2 min

+++ K_eff.m
@@ -3,7 +3,7 @@ function [D] = K_eff(D,i,Nb_SemiCycle)
 arguments (Input)
     D table
     i int32
-    Nb_SemiCycle int32
+    Nb_SemiCycle double
 end
 
 arguments (Output)
